refactor(skills): consolidate file structure — merge import/export into disk.ts, eliminate types.ts middleman
